+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # v10.1.1 - TBD
 
+## Enhancements
+
+- Add support for executing a script on a device via the device manager api [792](https://github.com/bugsnag/maze-runner/pull/792)
+
 ## Fixes
 
 - Downgrade Ruby in published Docker image to 3.3 to avoid repetitive frozen string literal warnings [788](https://github.com/bugsnag/maze-runner/pull/788)